Create Python_Implementation
This is the Python Implementation of the logic. The input is provided in IEEE 754, and output is converted to decimal back for comparision. This program tries to divide different types of input sets based on a change in the output exponent bit, or sign bit. It makes appropriate manipulations to the mantissa using XOR operations or shifting