Create positive_input
This module generates the output for positive inputs in the range [0,2). When inpt is 0, output is 0.5, and when input is 2, output is 1. Hence the exponent bit is always the same. However, the mantissa changes with each range of exponent input. To deal with this, the mantissa is right shifted as many times as the exponent is lower than 126. To add 0.5, an xor operation is done between the first bit after the shifted decimal point and 1.