Create negative_input
This declares the output for the input range (-2,0). There are two more cases in this situation: when the input is between -1 and -2 and when the input is between -1 and 0. This is because when input is in (-1,-2), The final output is in range (0, 0.25), which covers a lot of exponent ranges. Whereas, when the input is in the range (0, -1), the final output is in range (0.25, 0.5), which covers one exponent i.e. 125. Then the shifting occurs.